% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogmapr-package.R
\docType{package}
\name{cogmapr-package}
\alias{cogmapr}
\alias{cogmapr-package}
\title{cogmapr: Cognitive Mapping Tools Based on Coding of Textual Sources}
\description{
Functions for building cognitive maps based on qualitative data. Inputs are textual sources (articles, transcription of qualitative interviews of agents,...). These sources have been coded using relations and are linked to (i) a table describing the variables (or concepts) used for the coding and (ii) a table describing the sources (typology of agents, ...). Main outputs are Individual Cognitive Maps (ICM), Social Cognitive Maps (all sources or group of sources) and a list of quotes linked to relations. This package is linked to the work done during the PhD of Frederic M. Vanwindekens (CRA-W / UCL) hold the 13 of May 2014 at University of Louvain in collaboration with the Walloon Agricultural Research Centre (project MIMOSA, MOERMAN fund).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://frdvnw.gitlab.io/cogmapr/}
  \item Report bugs at \url{https://gitlab.com/FrdVnW/cogmapr/issues}
}

}
\author{
\strong{Maintainer}: Frédéric M. Vanwindekens \email{f.vanwindekens@cra.wallonie.be} (0000-0002-9117-7543)

Authors:
\itemize{
  \item Didier Stilmant \email{d.stimant@cra.wallonie.be} [thesis advisor]
  \item Philippe V. Baret \email{philippe.baret@uclouvain.be} [thesis advisor]
}

}
\keyword{internal}
