% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogmapr-indic.R
\name{GraphIndicatorsTable}
\alias{GraphIndicatorsTable}
\title{Table of graph indicators of a social cognitive map}
\usage{
GraphIndicatorsTable(df.graph.indic)
}
\arguments{
\item{df.graph.indic}{A data frame, as the output of the function 'GraphIndicators'}
}
\value{
A data frame of graph thery indicator, easier to read (long format)
}
\description{
Table of graph indicators of a social cognitive map
}
\details{
Table of graph indicators of a social cognitive map
}
\examples{
project_name <- "a_new_project"
main_path <- paste0(system.file("testdata", package = "cogmapr"), '/')
my.project <- ProjectCMap(main_path, project_name)

df.graph.indic <- GraphIndicators(my.project)
GraphIndicatorsTable(df.graph.indic)
}
